Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-float average(int a,int b,int c){
+float average(float a,float b,float c){
     float avg=(a+b+c)/3;
     return avg;
 }
